fix challenge form
