+++ src/app/course/challenge/course-challenge-create-edit/course-challenge-create-edit.component.scss
@@ -1,0 +1,5 @@
+@import "mixins";
+
+.question-islands {
+    @include island-grid();
+}
